Excel Parking Data analysis
The second sheet contains data analysis performed on three different data columns in our housing prices dataset. All three columns contained information on the number of parking and garage spaces for each home. An analysis was done to decide which columns to eliminate so only one data column was used in the machine learning model.